docs(demo): deleted demo package added demo on a per-package basis
