+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "postinstall": "npm run lerna:link && npm run lerna:bootstrap",
     "lerna:start": "lerna run start --stream",
-    "lerna:start:demo": "lerna run start --scope demo --stream",
+    "lerna:start:demo": "lerna run demo --stream --parallel",
     "lerna:build": "lerna run build",
     "lerna:bootstrap": "lerna bootstrap --no-ci",
     "lerna:link": "lerna link",
@@ -41,6 +41,7 @@
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.3",
     "lerna": "^3.22.0",
+    "parcel-bundler": "^1.12.4",
     "prettier": "^2.0.5",
     "rimraf": "^3.0.2",
     "rollup": "^2.11.2",
