--- packages/demo/CHANGELOG.md
@@ -1,32 +1,0 @@
-# Change Log
-
-All notable changes to this project will be documented in this file.
-See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
-
-## [1.0.4](https://github.com/dvelasquez/lighthouse-viewer/compare/demo@1.0.3...demo@1.0.4) (2020-06-12)
-
-**Note:** Version bump only for package demo
-
-
-
-
-
-## [1.0.3](https://github.com/dvelasquez/lighthouse-viewer/compare/demo@1.0.2...demo@1.0.3) (2020-05-27)
-
-**Note:** Version bump only for package demo
-
-
-
-
-
-## [1.0.2](https://github.com/dvelasquez/lighthouse-viewer/compare/demo@1.0.1...demo@1.0.2) (2020-05-24)
-
-**Note:** Version bump only for package demo
-
-
-
-
-
-## 1.0.1 (2020-05-24)
-
-**Note:** Version bump only for package demo
