--- packages/demo/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "demo",
-  "version": "1.0.4",
-  "private": true,
-  "description": "",
-  "main": "./dist/app.js",
-  "scripts": {
-    "prestart": "rimraf dist",
-    "start": "rollup -c -w",
-    "build": "rollup -c",
-    "test": "echo \"Error: no test specified\" && exit 0"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "react": "file:../react2-lighthouse-viewer/node_modules/react",
-    "react-dom": "^16.13.1",
-    "react2-lighthouse-viewer": "^0.0.1",
-    "vue": "file:../vue-lighthouse-viewer/node_modules/vue",
-    "vue-lighthouse-viewer": "^0.0.10"
-  },
-  "devDependencies": {
-    "@types/react": "file:../react2-lighthouse-viewer/node_modules/@types/react",
-    "@types/react-dom": "^16.9.8"
-  }
-}
