--- packages/demo/public/index.html
@@ -1,13 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <title>Demo Bundle</title>
-  </head>
-  <body>
-    <div id="vue-app"></div>
-    <br>
-    <div id="react-app"></div>
-    <script type="module" src="app.js"></script>
-  </body>
-</html>
