--- packages/demo/tsconfig.json
@@ -1,14 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "lib": ["ESNext", "ES6", "DOM"],
-        "strict": true,
-        "baseUrl": "src/",
-        "declaration": false,
-        "allowJs": false,
-        "jsx": "react"
-    },
-    "include": [
-        "./src/**/*"
-    ]
-}
