+++ packages/lighthouse-viewer/demo/index.html
@@ -1,0 +1,17 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="utf-8">
+    <title>Vanilla Demo Bundle</title>
+</head>
+<body>
+<main>
+    <div class="lh-root lh-vars">
+        <div id="html-template"></div>
+        <main class="lighthouse-viewer"></main>
+        <div id="lh-log"></div>
+    </div>
+</main>
+<script src="index.ts"></script>
+</body>
+</html>
