+++ packages/lighthouse-viewer/demo/report.json
