+++ packages/react2-lighthouse-viewer/demo/index.html
@@ -2,10 +2,10 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>Demo Bundle</title>
+    <title>React Demo Bundle</title>
 </head>
 <body>
 <div id="react-app"></div>
-<script type="module" src="app.js"></script>
+<script src="index.tsx"></script>
 </body>
 </html>
