+++ packages/react2-lighthouse-viewer/package.json
@@ -31,6 +31,7 @@
     "build:types": "cross-env NODE_ENV=production tsc --project tsconfig.d.json",
     "copy:types": "cpx \"src/types/**/*\" dist/types",
     "start": "rollup -c -w",
+    "demo": "parcel demo/index.html --port 8081 --open",
     "prebuild": "rimraf dist"
   },
   "dependencies": {
@@ -38,6 +39,8 @@
   },
   "devDependencies": {
     "@types/react": "^16.9.35",
-    "react": "^16.13.1"
+    "@types/react-dom": "^16.9.8",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1"
   }
 }
