+++ packages/vue-lighthouse-viewer/demo/index.html
@@ -2,10 +2,10 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>Demo Bundle</title>
+    <title>Vue Demo Bundle</title>
 </head>
 <body>
 <div id="vue-app"></div>
-<script type="module" src="app.js"></script>
+<script src="index.ts"></script>
 </body>
 </html>
