+++ packages/vue-lighthouse-viewer/package.json
@@ -31,6 +31,7 @@
     "build:types": "cross-env NODE_ENV=production tsc --project tsconfig.d.json",
     "copy:types": "cpx \"src/types/**/*\" dist/types",
     "start": "rollup -c -w",
+    "demo": "parcel demo/index.html --port 8082 --open",
     "prebuild": "rimraf dist"
   },
   "peerDependencies": {
@@ -41,5 +42,8 @@
   },
   "devDependencies": {
     "vue": "^2.6.11"
+  },
+  "alias": {
+    "vue": "./node_modules/vue/dist/vue.esm.js"
   }
 }
